+++ CPE200Lab1/CPE200Lab1/CPE200Lab1.csproj
@@ -61,6 +61,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RPNCalculatorEngine.cs" />
+    <Compile Include="SimpleCalculatorEngine.cs" />
     <EmbeddedResource Include="ExtendForm.resx">
       <DependentUpon>ExtendForm.cs</DependentUpon>
     </EmbeddedResource>
