+++ CPE200Lab1/CPE200Lab1/SimpleCalculatorEngine.cs
@@ -1,0 +1,60 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CPE200Lab1
+{
+	class SimpleCalculatorEngine: CalculatorEngine// lab4///
+	{
+		public string unaryCalculate(string operate, string operand, int maxOutputSize = 8)
+		{
+			switch (operate)
+			{
+				case "√":
+					{
+						double result;
+						string[] parts;
+						int remainLength;
+
+						result = Math.Sqrt(Convert.ToDouble(operand));
+						// split between integer part and fractional part
+						parts = result.ToString().Split('.');
+						// if integer part length is already break max output, return error
+						if (parts[0].Length > maxOutputSize)
+						{
+							return "E";
+						}
+						// calculate remaining space for fractional part.
+						remainLength = maxOutputSize - parts[0].Length - 1;
+						// trim the fractional part gracefully. =
+						return result.ToString("N" + remainLength);
+					}
+				case "1/x":
+					if (operand != "0")
+					{
+						double result;
+						string[] parts;
+						int remainLength;
+
+						result = (1.0 / Convert.ToDouble(operand));
+						// split between integer part and fractional part
+						parts = result.ToString().Split('.');
+						// if integer part length is already break max output, return error
+						if (parts[0].Length > maxOutputSize)
+						{
+							return "E";
+						}
+						// calculate remaining space for fractional part.
+						remainLength = maxOutputSize - parts[0].Length - 1;
+						// trim the fractional part gracefully. =
+						return result.ToString("N" + remainLength);
+					}
+					break;
+			}
+			return "E";
+		}
+
+	}
+}
